+++ lib/packages-dependencies.php
@@ -86,6 +86,7 @@ return array(
 		'wp-a11y',
 		'wp-api-fetch',
 		'wp-compose',
+		'wp-deprecated',
 		'wp-dom',
 		'wp-element',
 		'wp-hooks',